+++ src/GafferScene/RenderController.cpp
@@ -319,8 +319,8 @@ class RenderController::SceneGraph
 			AttributesComponent = 4,
 			ObjectComponent = 8,
 			ChildNamesComponent = 16,
-			ExpansionComponent = 32,
-			AllComponents = BoundComponent | TransformComponent | AttributesComponent | ObjectComponent | ChildNamesComponent | ExpansionComponent,
+			VisibleSetComponent = 32,
+			AllComponents = BoundComponent | TransformComponent | AttributesComponent | ObjectComponent | ChildNamesComponent | VisibleSetComponent,
 		};
 
 		// Constructs the root of the scene graph.
@@ -437,6 +437,20 @@ class RenderController::SceneGraph
 
 			clean( TransformComponent );
 
+			// VisibleSet
+
+			const auto currentDrawMode = m_drawMode;
+			if( ( m_dirtyComponents & VisibleSetComponent ) && updateVisibleSet( path, controller->m_visibleSet, controller->m_minimumExpansionDepth ) )
+			{
+				m_changedComponents |= VisibleSetComponent;
+
+				// An object update is only required on change of draw mode for this location.
+				if( currentDrawMode != m_drawMode )
+				{
+					m_dirtyComponents |= ObjectComponent;
+				}
+			}
+
 			// Object
 
 			if( ( m_dirtyComponents & ObjectComponent ) && updateObject( controller->m_scene->objectPlug(), type, controller->m_renderer.get(), controller->m_globals.get(), controller->m_scene.get(), controller->m_lightLinks.get(), controller->m_motionBlurOptions ) )
@@ -523,29 +537,25 @@ class RenderController::SceneGraph
 
 			clean( ChildNamesComponent );
 
-			// Expansion
-
-			if( ( m_dirtyComponents & ExpansionComponent ) && updateExpansion( path, controller->m_visibleSet, controller->m_minimumExpansionDepth ) )
-			{
-				m_changedComponents |= ExpansionComponent;
-			}
-
 			bool newBound = false;
 			if(
-				( m_changedComponents & ( ExpansionComponent | ChildNamesComponent ) ) ||
+				( m_changedComponents & ( VisibleSetComponent | ChildNamesComponent ) ) ||
 				( m_dirtyComponents & BoundComponent )
 			)
 			{
 				// Create bounding box if needed
 				Box3f bound;
-				if( !m_expanded && m_children.size() )
+				if( ( m_drawMode == VisibleSet::Visibility::Visible && !m_descendantsVisible && m_children.size() ) || m_drawMode == VisibleSet::Visibility::ExcludedBounds )
 				{
 					bound = controller->m_scene->boundPlug()->getValue();
 				}
 
 				if( !bound.isEmpty() )
 				{
-					IECoreScenePreview::PlaceholderPtr placeholder = new IECoreScenePreview::Placeholder( bound );
+					IECoreScenePreview::PlaceholderPtr placeholder = new IECoreScenePreview::Placeholder(
+						bound,
+						m_drawMode == VisibleSet::Visibility::ExcludedBounds ? IECoreScenePreview::Placeholder::Mode::Excluded : IECoreScenePreview::Placeholder::Mode::Default
+					);
 
 					std::string boundName;
 					ScenePlug::pathToString( path, boundName );
@@ -583,7 +593,7 @@ class RenderController::SceneGraph
 				}
 			}
 
-			clean( ExpansionComponent | BoundComponent );
+			clean( VisibleSetComponent | BoundComponent );
 
 			m_cleared = false;
 
@@ -594,7 +604,7 @@ class RenderController::SceneGraph
 
 		bool expanded() const
 		{
-			return m_expanded;
+			return m_descendantsVisible;
 		}
 
 		const std::vector<std::unique_ptr<SceneGraph>> &children()
@@ -620,7 +630,8 @@ class RenderController::SceneGraph
 			clearObject();
 			m_attributesHash = m_lightLinksHash = m_transformHash = m_childNamesHash = IECore::MurmurHash();
 			m_cleared = true;
-			m_expanded = false;
+			m_descendantsVisible = false;
+			m_drawMode = VisibleSet::Visibility::None;
 			m_boundInterface = nullptr;
 			m_dirtyComponents = AllComponents;
 		}
@@ -761,7 +772,7 @@ class RenderController::SceneGraph
 		bool updateObject( const ObjectPlug *objectPlug, Type type, IECoreScenePreview::Renderer *renderer, const IECore::CompoundObject *globals, const ScenePlug *scene, LightLinks *lightLinks, const MotionBlurOptions &motionBlurOptions )
 		{
 			const bool hadObjectInterface = static_cast<bool>( m_objectInterface );
-			if( type == NoType )
+			if( type == NoType || m_drawMode != VisibleSet::Visibility::Visible )
 			{
 				clearObject();
 				return hadObjectInterface;
@@ -957,17 +968,17 @@ class RenderController::SceneGraph
 			m_objectHash = MurmurHash();
 		}
 
-		bool updateExpansion( const ScenePlug::ScenePath &path, const GafferScene::VisibleSet &visibleSet, size_t minimumExpansionDepth )
+		bool updateVisibleSet( const ScenePlug::ScenePath &path, const GafferScene::VisibleSet &visibleSet, size_t minimumExpansionDepth )
 		{
-			/// \todo We also include DescendantMatch here to ensure the ancestors of visible paths are also visible,
-			/// ideally we'd avoid expanding those ancestors and instead render only the ExactMatch.
-			const bool expanded = visibleSet.match( path, minimumExpansionDepth ) & ( PathMatcher::ExactMatch | PathMatcher::DescendantMatch );
+			const auto visibility = visibleSet.visibility( path, minimumExpansionDepth );
 
-			if( expanded == m_expanded )
+			if( visibility.descendantsVisible == m_descendantsVisible && visibility.drawMode == m_drawMode )
 			{
 				return false;
 			}
-			m_expanded = expanded;
+			m_descendantsVisible = visibility.descendantsVisible;
+			m_drawMode = visibility.drawMode;
+
 			return true;
 		}
 
@@ -1106,7 +1117,8 @@ class RenderController::SceneGraph
 		std::vector<std::unique_ptr<SceneGraph>> m_children;
 
 		IECoreScenePreview::Renderer::ObjectInterfacePtr m_boundInterface;
-		bool m_expanded;
+		bool m_descendantsVisible;
+		VisibleSet::Visibility::DrawMode m_drawMode;
 
 		// Tracks work which needs to be done on
 		// the next call to `update()`.
@@ -1393,7 +1405,7 @@ void RenderController::setVisibleSet( const GafferScene::VisibleSet &visibleSet 
 	cancelBackgroundTask();
 
 	m_visibleSet = visibleSet;
-	dirtySceneGraphs( SceneGraph::ExpansionComponent );
+	dirtySceneGraphs( SceneGraph::VisibleSetComponent );
 	requestUpdate();
 }
 
@@ -1412,7 +1424,7 @@ void RenderController::setMinimumExpansionDepth( size_t depth )
 	cancelBackgroundTask();
 
 	m_minimumExpansionDepth = depth;
-	dirtySceneGraphs( SceneGraph::ExpansionComponent );
+	dirtySceneGraphs( SceneGraph::VisibleSetComponent );
 	requestUpdate();
 }